requirements.txt
asgiref==3.7.2
Django==4.2.6
django-admin-soft-dashboard==1.0.12
django-admin-volt==1.0.10
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
greenlet==3.0.0
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
Pillow==10.0.1
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
XlsxWriter==3.1.9
